Public

Public Documentation

Documentation for CovarianceEstimation.jl's public interface.

See Internal Documentation for internal package docs.

Contents

Public Documentation

Contents

Index

Public Interface

Index

CovarianceEstimation.AnalyticalNonlinearShrinkage

CovarianceEstimation.CommonCovariance

CovarianceEstimation.ConstantCorrelation

CovarianceEstimation.CovarianceEstimator

CovarianceEstimation.DiagonalCommonVariance

CovarianceEstimation.DiagonalUnequalVariance

CovarianceEstimation.DiagonalUnitVariance

CovarianceEstimation.LinearShrinkage

CovarianceEstimation.PerfectPositiveCorrelation

CovarianceEstimation.Simple

Statistics.cov

Public Interface

Statistics.cov — Function.

cov(sc::Simple, x::AbstractVector, y::AbstractVector)

Compute the covariance of the vectors x and y using formula $\frac{1}{n}\sum_{i=1}^n (x_i-\bar x) (y_i-\bar y)^*$ where $*$ denotes complex conjugate. If sc.corrected is true then the fraction $\frac{1}{n}$ is replaced with $\frac{1}{n-1}$.

source

cov(sc::Simple, X::AbstractMatrix; dims=1, mean=nothing)

Compute the sample variance of X. The sum is scaled with the number of observations n if sc.corrected is false and with n-1 otherwise. If dims=1 the rows are assumed to be observations and the columns the features. If dims=2 the opposite is assumed. The keyword mean can be

* `nothing` (default) in which case the mean is estimated then subtracted
from the data `X`,
* `zero` in which case the data is assumed to have already been centred,
* a given vector of appropriate dimensions in which case that provided mean
is subtracted from the data `X`.

source

cov(sc::Simple, X::AbstractVector; mean=nothing)

Compute the sample variance of X. The sum is scaled with the number of observations n if sc.corrected is false and with n-1 otherwise. The keyword mean can be

* `nothing` (default) in which case the mean is estimated then subtracted
from the data `X`,
* `zero` in which case the data is assumed to have already been centred,
* a given number in which case that provided mean is subtracted from the
data `X`.

source

cov(lse::LinearShrinkage, X; dims=1)

Linear shrinkage covariance estimator for matrix X along dimension dims. Computed using the method described by lse.

source

cov(ans::AnalyticalNonlinearShrinkage, X; dims=1, mean=nothing)

Nonlinear covariance estimator derived from the sample covariance estimator S and its eigenvalue decomposition (which can be given through decomp). See Ledoit and Wolf's paper http://www.econ.uzh.ch/static/wp/econwp264.pdf The keyword mean can be nothing (centering via estimated mean), zero (no centering) or a provided vector. In the first case, a rank-1 modification is applied and therefore the effective sample size is decreased by one (see analytical_nonlinear_shrinkage). In the latter two case the mean cannot have been estimated on the data (otherwise the effective sample size will be 1 larger than it should be resulting in numerical instabilities). If you are unsure, use either nothing or provide an explicit (non-estimated) vector (possibly a zero vector) and avoid the use of mean=0.

Time complexity (including formation of S)

(p<n): O(np^2 + n^2) with moderate constant

(p>n): O(p^3) with low constant (dominated by eigendecomposition of S)

source

CovarianceEstimation.CovarianceEstimator — Type.

CovarianceEstimator

Basic type for all covariance estimators.

source

CovarianceEstimation.Simple — Type.

Simple(corrected::Bool, weights::Union{AbstractWeights, Nothing} = nothing)

Simple covariance estimator (scaled by n-1 if corrected and n otherwise where n is the number of samples). Optionally supports weights (see http://juliastats.github.io/StatsBase.jl/stable/cov.html).

source

CovarianceEstimation.LinearShrinkage — Type.

LinearShrinkage(target, shrinkage; corrected=false)

Linear shrinkage estimator described by equation $(1 - \lambda) S + \lambda F$ where $S$ is standard covariance matrix, $F$ is shrinkage target described by argument target and $\lambda$ is a shrinkage parameter, either given explicitly in shrinkage or automatically determined according to one of the supported methods.

The corrected estimator is used if corrected is true.

source

CovarianceEstimation.DiagonalUnitVariance — Type.

DiagonalUnitVariance

Target for linear shrinkage: unit matrix. A subtype of LinearShrinkageTarget where

$F_{ij}=1$ if $i=j$ and

$F_{ij}=0$ otherwise

source

CovarianceEstimation.DiagonalCommonVariance — Type.

DiagonalCommonVariance

Target for linear shrinkage: unit matrix multiplied by average variance of variables. A subtype of LinearShrinkageTarget where

$F_{ij}=v$ if $i=j$ with $v=\mathrm{tr}(S)/p$ and

$F_{ij}=0$ otherwise

source

CovarianceEstimation.DiagonalUnequalVariance — Type.

DiagonalUnequalVariance

Target for linear shrinkage: diagonal of covariance matrix. A subtype of LinearShrinkageTarget where

$F_{ij}=s_{ij}$ if $i=j$ and

$F_{ij}=0$ otherwise

source

CovarianceEstimation.CommonCovariance — Type.

CommonCovariance

Target for linear shrinkage: see target_C. A subtype LinearShrinkageTarget where

$F_{ij}=v$ if $i=j$ with $v=\mathrm{tr}(S)/p$ and

$F_{ij}=c$ with $c=\sum_{i\neq j} S_{ij}/(p(p-1))$ otherwise

source

CovarianceEstimation.PerfectPositiveCorrelation — Type.

PerfectPositiveCorrelation

Target for linear shrinkage: see target_E. A subtype of LinearShrinkageTarget where

$F_{ij}=S_{ij}$ if $i=j$ and

$F_{ij}=\sqrt{S_{ii}S_{jj}}$ otherwise

source

CovarianceEstimation.ConstantCorrelation — Type.

ConstantCorrelation

Target for linear shrinkage: see target_F. A subtype of LinearShrinkageTarget where

$F_{ij}=S_{ij}$ if $i=j$ and

$F_{ij}=\overline{r}\sqrt{S_{ii}S_{jj}}$ otherwise where $\overline{r}$ is the average sample correlation

source

CovarianceEstimation.AnalyticalNonlinearShrinkage — Type.

AnalyticalNonlinearShrinkage

Analytical nonlinear shrinkage estimator. See docs for analytical_nonlinear_shrinkage for details.

source