Internals

Internal Documentation

Documentation for CovarianceEstimation.jl's internal methods.

See Public Documentation for public package docs.

Contents

Internal Documentation

Contents

Index

Internals

Index

CovarianceEstimation.analytical_nonlinear_shrinkage

CovarianceEstimation.linear_shrinkage

CovarianceEstimation.rescale

CovarianceEstimation.square

CovarianceEstimation.sum_fij

CovarianceEstimation.sumij

CovarianceEstimation.sumij2

CovarianceEstimation.uccov

Internals

CovarianceEstimation.rescale — Function.

rescale(M, d)

Internal function to scale the rows and the columns of a square matrix M according to the elements in d. This is useful when dealing with the standardised data matrix which can be written Xs=Xc*D where Xc is the centered data matrix and D=Diagonal(d) so that its simple covariance is D*S*D where S is the simple covariance of Xc. Such D*M*D terms appear often in the computations of optimal shrinkage λ.

Space complexity: $O(p^2)$

Time complexity: $O(p^2)$

source

CovarianceEstimation.uccov — Function.

uccov(X)

Internal function to compute X*X'/n where n is the number of rows of X. This corresponds to the uncorrected covariance of X if X is centered. This operation appears often in the computations of optimal shrinkage λ.

Space complexity: $O(p^2)$

Time complexity: $O(2np^2)$

source

CovarianceEstimation.sumij — Function.

sumij(S)

Internal function to compute the sum of elements of a square matrix S. A keyword with_diag can be passed to indicate whether to include or not the diagonal of S in the sum. Both cases happen often in the computations of optimal shrinkage λ.

Space complexity: $O(1)$

Time complexity: $O(p^2)$

source

CovarianceEstimation.square — Function.

square(x)

Internal function to compute the square of a real number x. Defined here so it can be used in the other internal function sumij2.

source

CovarianceEstimation.sumij2 — Function.

sumij2(S)

Internal function identical to sumij except that it passes the function square to the sum so that it is the sum of the elements of S squared which is computed. This is significantly more efficient than using sumij(S.^2) for large matrices as it allocates very little.

Space complexity: $O(1)$

Time complexity: $O(2p^2)$

source

CovarianceEstimation.sum_fij — Function.

sum_fij(Xc, S, n, κ)

Internal function corresponding to $∑_{i≂̸j}f_{ij}$ that appears in http://strimmerlab.org/publications/journals/shrinkcov2005.pdf p.11.

Space complexity: $O(np + 2p^2)$

Time complexity: $O(2np^2)$

source

CovarianceEstimation.linear_shrinkage — Function.

linear_shrinkage(target, Xc, S, λ, n, p, corrected)

Performs linear shrinkage with target of type target for data matrix Xc of size n by p with covariance matrix S and shrinkage parameter λ. Calculates corrected covariance if corrected is true.

source

CovarianceEstimation.analytical_nonlinear_shrinkage — Function.

analytical_nonlinear_shrinkage(S, n, p; decomp)

Internal implementation of the analytical nonlinear shrinkage. The implementation is inspired from the Matlab code given in section C of Olivier Ledoit and Michael Wolf's paper "Analytical Nonlinear Shrinkage of Large-Dimensional Covariance Matrices". (Nov 2018) http://www.econ.uzh.ch/static/wp/econwp264.pdf

source