Public

Public Documentation

Documentation for CovarianceEstimation.jl's public interface.

See Internal Documentation for internal package docs.

Contents

Public Documentation

Contents

Index

Public Interface

Index

CovarianceEstimation.AnalyticalNonlinearShrinkage

CovarianceEstimation.CommonCovariance

CovarianceEstimation.ConstantCorrelation

CovarianceEstimation.CovarianceEstimator

CovarianceEstimation.DiagonalCommonVariance

CovarianceEstimation.DiagonalUnequalVariance

CovarianceEstimation.DiagonalUnitVariance

CovarianceEstimation.LinearShrinkageEstimator

CovarianceEstimation.PerfectPositiveCorrelation

CovarianceEstimation.Simple

Statistics.cov

Public Interface

Statistics.cov — Function.

cov(x::AbstractVector, sc::Simple)

Compute the sample variance of the vector x. The sum is scaled with n where n = length(x) if sc.corrected is false and with n-1 otherwise.

source

cov(x::AbstractVector, y, sc::Simple)

Compute the covariance of the vectors x and y using formula $\frac{1}{n}\sum_{i=1}^n (x_i-\bar x) (y_i-\bar y)^*$ where $*$ denotes complex conjugate. If sc.corrected is true then the fraction $\frac{1}{n}$ is replaced with $\frac{1}{n-1}$.

source

cov(X::AbstractMatrix, sc::Simple; dims::Int=1)

Compute the covariance matrix associated with X along dimension dims. The sum is scaled with n where n = length(x) if sc.corrected is false and with n-1 otherwise.

source

cov(X, sc::Simple{<:AbstractWeights}; dims=1)

Compute the weighted covariance matrix.

source

cov(X, lse::LinearShrinkageEstimator; dims=1)

Linear shrinkage covariance estimator for matrix X along dimension dims. Computed using the method described by lse.

source

cov(X, ans::AnalyticalNonlinearShrinkage; dims=1)

source

CovarianceEstimation.CovarianceEstimator — Type.

CovarianceEstimator

Basic type for all covariance estimators.

source

CovarianceEstimation.Simple — Type.

Simple(corrected::Bool, weights::Union{AbstractWeights, Nothing} = nothing)

Simple covariance estimator (scaled by n-1 if corrected and n otherwise where n is the number of samples). Optionally supports weights (see http://juliastats.github.io/StatsBase.jl/stable/cov.html).

source

CovarianceEstimation.LinearShrinkageEstimator — Type.

LinearShrinkageEstimator(target, shrinkage; corrected = false)

Linear shrinkage estimator described by equation $(1 - \lambda) S + \lambda F$ where $S$ is standard covariance matrix, $F$ is shrinkage target described by argument target and $\lambda$ is a shrinkage parameter, either given explicitly in shrinkage or automatically determined according to one of the supported methods.

The corrected estimator is used if corrected is true.

source

CovarianceEstimation.DiagonalUnitVariance — Type.

DiagonalUnitVariance

Target for linear shrinkage: unit matrix. A subtype of LinearShrinkageTarget where

$F_{ij}=1$ if $i=j$ and

$F_{ij}=0$ otherwise

source

CovarianceEstimation.DiagonalCommonVariance — Type.

DiagonalCommonVariance

Target for linear shrinkage: unit matrix multiplied by average variance of variables. A subtype of LinearShrinkageTarget where

$F_{ij}=v$ if $i=j$ with $v=\mathrm{tr}(S)/p$ and

$F_{ij}=0$ otherwise

source

CovarianceEstimation.DiagonalUnequalVariance — Type.

DiagonalUnequalVariance

Target for linear shrinkage: diagonal of covariance matrix. A subtype of LinearShrinkageTarget where

$F_{ij}=s_{ij}$ if $i=j$ and

$F_{ij}=0$ otherwise

source

CovarianceEstimation.CommonCovariance — Type.

CommonCovariance

Target for linear shrinkage: see target_C. A subtype LinearShrinkageTarget where

$F_{ij}=v$ if $i=j$ with $v=\mathrm{tr}(S)/p$ and

$F_{ij}=c$ with $c=\sum_{i\neq j} S_{ij}/(p(p-1))$ otherwise

source

CovarianceEstimation.PerfectPositiveCorrelation — Type.

PerfectPositiveCorrelation

Target for linear shrinkage: see target_E. A subtype of LinearShrinkageTarget where

$F_{ij}=S_{ij}$ if $i=j$ and

$F_{ij}=\sqrt{S_{ii}S_{jj}}$ otherwise

source

CovarianceEstimation.ConstantCorrelation — Type.

ConstantCorrelation

Target for linear shrinkage: see target_F. A subtype of LinearShrinkageTarget where

$F_{ij}=S_{ij}$ if $i=j$ and

$F_{ij}=\overline{r}\sqrt{S_{ii}S_{jj}}$ otherwise where $\overline{r}$ is the average sample correlation

source

CovarianceEstimation.AnalyticalNonlinearShrinkage — Type.

AnalyticalNonlinearShrinkage

Analytical nonlinear shrinkage estimator. See docs for analytical_nonlinear_shrinkage for details.

source